backend/AI_backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
networkx
matplotlib
groq